requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alabaster==1.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
app-model==0.4.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
babel==2.18.0
bermuda==0.1.6
build==1.4.0
cachey==0.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
dask==2026.1.2
debugpy==1.8.20
decorator==5.2.1
docstring_parser==0.17.0
docutils==0.22.4
donfig==0.8.1.post1
executing==2.2.1
filelock==3.25.0
flexcache==0.3
flexparser==0.4
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2026.2.0
google-crc32c==1.8.0
HeapDict==1.0.1
hsluv==5.0.4
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
in-n-out==0.2.1
iniconfig==2.3.0
iopath==0.1.10
ipykernel==6.31.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
locket==1.0.0
magicgui==0.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
napari==0.6.6
napari-console==0.1.4
napari-plugin-engine==0.2.1
napari-plugin-manager==0.1.10
napari-svg==0.2.1
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.3.3
npe2==0.8.1
numba==0.63.1
numcodecs==0.16.5
numpy==2.3.5
numpydoc==1.10.0
omegaconf==2.3.0
packaging==26.0
pandas==3.0.0
parso==0.8.6
partd==1.4.2
pexpect==4.9.0
pillow==12.1.1
Pint==0.25.2
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.9.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyconify==0.2.1
pydantic==2.12.5
pydantic-compat==0.1.2
pydantic-extra-types==2.11.0
pydantic_core==2.41.5
pydicom==3.0.1
Pygments==2.19.2
pylibjpeg==2.1.0
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.5.0
PyOpenGL==3.1.10
pyproject_hooks==1.2.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gdcm==3.2.2
PyYAML==6.0.3
pyzmq==27.1.0
qtconsole==5.7.1
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rich==14.3.2
roman-numerals==4.1.0
rpds-py==0.30.0
sam2==1.1.0
scikit-image==0.26.0
scipy==1.17.0
setuptools==82.0.0
shellingham==1.5.4
shiboken6==6.7.2
six==1.17.0
snowballstemmer==3.0.1
Sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
superqt==0.7.8
sympy==1.14.0
tifffile==2026.1.28
tomli_w==1.2.0
toolz==1.1.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
triangle==20250106
typer==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vispy==0.15.2
wcwidth==0.6.0
wrapt==2.1.1
yarl==1.22.0
zarr==3.1.5
